% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_model_functions.R
\name{brent_cox_scheffe_gaussian}
\alias{brent_cox_scheffe_gaussian}
\title{Minimizes the efficiency function for Scheffé model using Brent's method.
R wrapper for BrentCoxScheffeGaussian in C++.
TODO: write doc}
\usage{
brent_cox_scheffe_gaussian(
  X,
  j,
  i,
  order,
  opt_crit,
  lwr_bnd = 0,
  uppr_bnd = 1,
  tol = 1e-04
)
}
\description{
Minimizes the efficiency function for Scheffé model using Brent's method.
R wrapper for BrentCoxScheffeGaussian in C++.
TODO: write doc
}
