% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_model_functions.R
\name{brent_global_cox_scheffe_gaussian}
\alias{brent_global_cox_scheffe_gaussian}
\title{R wrapper for BrentGloCoxScheffeGaussian in C++.
Minimizes the efficiency function for Scheffé model using Brent's method for global optima.
Needs a bound for the second derivative.
TODO: write doc}
\usage{
brent_global_cox_scheffe_gaussian(
  X,
  j,
  i,
  order,
  opt_crit,
  lower = 0,
  upper = 1,
  initial_guess = 0.5,
  hessian_bound = 1e+05,
  abs_err_tol = 1e-04,
  tol = 1e-04
)
}
\description{
R wrapper for BrentGloCoxScheffeGaussian in C++.
Minimizes the efficiency function for Scheffé model using Brent's method for global optima.
Needs a bound for the second derivative.
TODO: write doc
}
