% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnl_model_functions.R
\name{create_random_beta}
\alias{create_random_beta}
\title{Creation of a random parameter vector for MNL model}
\usage{
create_random_beta(q, seed = NULL)
}
\arguments{
\item{q}{integer specifying the number of ingredients}
}
\value{
Returns a list in which the first element of the list is
a numerical vector with the parameters and the second element is a matrix with
the indices as in Ruseckaite, et al - Bayesian D-optimal choice designs for mixtures (2017)
}
\description{
\code{create_random_beta} creates a random parameter vector
}
\examples{
create_random_beta(3)
}
