% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_model_functions.R
\name{efficiency_cox_scheffe_gaussian}
\alias{efficiency_cox_scheffe_gaussian}
\title{Computes efficiency criterion of a design matrix X but where the j-th ingredient in i-th observation is changed to theta.
TODO: write doc}
\usage{
efficiency_cox_scheffe_gaussian(theta, X, j, i, order, opt_crit)
}
\description{
Computes efficiency criterion of a design matrix X but where the j-th ingredient in i-th observation is changed to theta.
TODO: write doc
}
