% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_model_functions.R
\name{gaussian_create_moment_matrix_numerically}
\alias{gaussian_create_moment_matrix_numerically}
\title{TODO: write doc
Function created for testing. There's really no reason to use this function.}
\usage{
gaussian_create_moment_matrix_numerically(
  q,
  n_pv = 0,
  order = 3,
  n_points = 5000,
  pv_bounds = c(-1, 1),
  seed = NULL
)
}
\description{
TODO: write doc
Function created for testing. There's really no reason to use this function.
}
