% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_model_functions.R
\name{gaussian_mixture_coord_exch}
\alias{gaussian_mixture_coord_exch}
\title{Coordinate exchange algorithm for a mixture model assuming Gaussian iid errors.}
\usage{
gaussian_mixture_coord_exch(
  n_runs = NULL,
  q = NULL,
  n_random_starts = 100,
  X = NULL,
  order = 1,
  opt_method = "B",
  max_it = 10,
  tol = 1e-04,
  n_cox_points = NULL,
  plot_designs = F,
  verbose = 1,
  opt_crit = 0,
  seed = NULL,
  n_cores = 1,
  n_pv = 0
)
}
\arguments{
\item{n_runs}{number of runs}

\item{q}{number of ingredient proportions}

\item{n_random_starts}{number or random starts. Defaults to 100.}

\item{X}{User supplied design matrix. Must be of size (n_runs, q) where:
n_runs is the number of runs
q is the number of ingredient proportions.}

\item{order}{Order of the Scheffé model (1, 2, or 3).}

\item{opt_method}{Optimization method in each step of the coordinate exchange algorithm.
It can be "B" (Brent's algorithm) or "D" (discretization of Cox direction)}

\item{max_it}{integer for maximum number of iterations that the coordinate exchange algorithm will do}

\item{tol}{A positive error tolerance in Brent's method.}

\item{n_cox_points}{number of points to use in the discretization of Cox direction}

\item{plot_designs}{boolean. If TRUE, shows a plot of the initial and the final design. Only works if q is 3.}

\item{verbose}{level of verbosity.}

\item{opt_crit}{optimality criterion: D-optimality ("D" or 0) or I-optimality ("I" or 1)}

\item{seed}{Seed for reproducibility}

\item{n_cores}{Number of cores for parallel processing}

\item{n_pv}{Number of process variables.}
}
\value{
List with the following elements: \itemize{
\item \code{X_orig}: The original design. Matrix of size (n_runs, q).
\item \code{X}: The optimized design. Matrix of size (n_runs, q).
\item \code{opt_crit_value_orig}: efficiency of the original design.
\item \code{opt_crit_value}: efficiency of the optimized design.
\item \code{n_iter}: Number of iterations performed.
\item \code{efficiency_value_per_iteration}:
\item \code{opt_crit}: The optimality criterion used.
\item \code{q}: Number of mixture ingredients.
\item \code{n_pv}: Number of process variables.
\item \code{seed}: seed used to generate the final design. If a design was used as input by the user, this will be NA.
}
}
\description{
Performs the coordinate exchange algorithm for a Scheffé mixture model. It can have many different random starts, or the coordinate exchange algorithm can be performed in a user-supplied matrix.
}
