% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_model_functions.R
\name{gaussian_plot_result}
\alias{gaussian_plot_result}
\title{Plots the result of the list that results from the \code{gaussian_mixture_coord_exch()} function.}
\usage{
gaussian_plot_result(res_alg)
}
\arguments{
\item{res_alg}{List resulting from a call to the \code{gaussian_mixture_coord_exch()} function.}
}
\description{
Plots the result of the list that results from the \code{gaussian_mixture_coord_exch()} function.
}
\details{
Only works for \code{q = 3}.
}
