% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnl_model_functions.R
\name{get_opt_crit_value_MNL}
\alias{get_opt_crit_value_MNL}
\title{Function that computes optimality criterion value for MNL model}
\usage{
get_opt_crit_value_MNL(X, beta, order, opt_crit = 0)
}
\arguments{
\item{X}{3 dimensional array with dimensions (q, J, S) where:
q is the number of ingredient proportions,
J is the number of alternatives within a choice set,
S is the number of choice sets.}

\item{beta}{numeric vector containing the parameters. Should be of length (q^3 + 5*q)/6. Must extend to Bayesian case and have the option of it to be a matrix.}

\item{opt_crit}{optimality criterion: 0 is D-optimality and 1 is I-optimality}
}
\value{
Returns the value of the optimality criterion for this particular design and this beta vector
}
\description{
\code{get_opt_crit_value_MNL} computes optimality criterion value for MNL model
}
