% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{mixture_coord_ex}
\alias{mixture_coord_ex}
\title{TODO: write doc}
\usage{
mixture_coord_ex(
  n_random_starts = 100,
  X = NULL,
  order = NULL,
  n_runs = NULL,
  q = NULL,
  J = NULL,
  S = NULL,
  beta = NULL,
  model = "Gaussian",
  opt_method = "B",
  max_it = 10,
  tol = 1e-04,
  n_cox_points = NULL,
  plot_designs = F,
  verbose = 1,
  opt_crit = 0,
  seed = NULL,
  n_cores = 1
)
}
\description{
TODO: write doc
}
