% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_model_functions.R
\name{mixture_coord_ex_gaussian}
\alias{mixture_coord_ex_gaussian}
\title{TODO: write doc}
\usage{
mixture_coord_ex_gaussian(
  n_runs = NULL,
  q = NULL,
  n_random_starts = 100,
  X = NULL,
  order = 1,
  n_cox_points = 30,
  max_it = 10,
  plot_designs = F,
  verbose = 1,
  opt_crit = 0,
  seed = NULL,
  n_cores = 1
)
}
\description{
TODO: write doc
}
