% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnl_model_functions.R
\name{mixture_coord_ex_mnl}
\alias{mixture_coord_ex_mnl}
\title{Coordinate exchange algorithm for a Multinomial Logit Scheffé model.}
\usage{
mixture_coord_ex_mnl(
  X,
  beta,
  n_cox_points = 100,
  max_it = 50,
  plot_designs = F,
  verbose = 1
)
}
\arguments{
\item{X}{3 dimensional array with dimensions (q, J, S) where:
q is the number of ingredient proportions,
J is the number of alternatives within a choice set,
S is the number of choice sets.}

\item{beta}{numeric vector containing the parameters. Should be of length (q^3 + 5*q)/6}

\item{n_cox_points}{integer representing the umber of points to use in the discretization of Cox direction}

\item{max_it}{integer for maximum number of iterations that the coordinate exchange algorithm will do}

\item{plot_designs}{boolean. If TRUE, shows a plot of the initial and the final design. Only works if q is 3.}

\item{verbose}{level of verbosity. 6 levels. See below for details.}
}
\value{
list with 5 elements. See below for details.

Verbosity levels: each level prints the previous plus additional things:
\enumerate{
\item Print the log D efficiency in each iteration and a final summary
\item Print the values of k, s, i, and log D efficiency in each subiteration
\item Print the resulting X after each iteration, i.e., after each complete pass on the data
\item Print log D efficiency for each point in the Cox direction discretization
\item Print the resulting X and information matrix after each subiteration
\item Print the resulting X or each point in the Cox direction discretization
}

Return list has 5 elements:
\itemize{
\item X_orig: The original design. Array with dimensions (q, J, S).
\item X: The optimized design. Array with dimensions (q, J, S).
\item d_eff_orig: log D-efficiency of the original design.
\item d_eff: log D-efficiency of the optimized design.
\item n_iter: Number of iterations performed.
}
}
\description{
\code{mixture_coord_ex_mnl} Performs the coordinate exchange algorithm for a Multinomial Logit Scheffé model.
}
