% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnl_model_functions.R
\name{mnl_create_random_initial_design}
\alias{mnl_create_random_initial_design}
\title{Creation of a random initial design for MNL model}
\usage{
mnl_create_random_initial_design(
  q,
  J,
  S,
  n_pv = 0,
  pv_bounds = NULL,
  seed = NULL
)
}
\arguments{
\item{q}{integer specifying the number of ingredients}

\item{J}{integer specifying the number of alternatives within each choice set}

\item{S}{integer specifying the number of choice sets}

\item{seed}{integer used for reproducibility}
}
\value{
3-dimensional array of dimensions (q, J, S)
}
\description{
\code{mnl_create_random_initial_design} creates a random initial design for MNL model of the specified dimensions
}
\examples{
mnl_create_random_initial_design(3, 5, 4, seed = 2020)
}
