% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnl_model_functions.R
\name{mnl_get_number_of_parameters}
\alias{mnl_get_number_of_parameters}
\title{Get number of parameters.}
\usage{
mnl_get_number_of_parameters(q, n_pv, order)
}
\arguments{
\item{q}{integer specifying the number of mixture ingredients}

\item{n_pv}{Number of process variables}

\item{order}{integer corresponding to a Scheffé model order (1, 2, 3 if no process variables are used; 2 if process variables are used).}
}
\description{
Returns an integer with the number of parameters depending on \code{q}, \code{n_pv}, and \code{order}.
}
