% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnl_model_functions.R
\name{mnl_get_Ps}
\alias{mnl_get_Ps}
\title{TODO: write doc
Wrapper function for getPsMNL()}
\usage{
mnl_get_Ps(X, beta, s, order = 3, transform_beta = T)
}
\description{
TODO: write doc
Wrapper function for getPsMNL()
}
