% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnl_model_functions.R
\name{mnl_get_Xs}
\alias{mnl_get_Xs}
\title{Compute design matrix of a specific choice set using the MNL model.}
\usage{
mnl_get_Xs(X, s, order)
}
\arguments{
\item{X}{3-dimensional design cube/array of size \code{(q, J, S)}.}

\item{s}{integer corresponding to a choice set in 1 to S.}

\item{order}{integer corresponding to a Scheffé model order (1, 2, 3).}
}
\value{
Matrix of dimension (J, m-1) that corresponds to the design matrix of choice set s.
}
\description{
Compute the design matrix of choice set \code{s} using the MNL model.
}
