% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnl_model_functions.R
\name{mnl_plot_result_3d}
\alias{mnl_plot_result_3d}
\title{TODO: write doc}
\usage{
mnl_plot_result_3d(
  res_alg,
  design = "final",
  phi = 40,
  theta = 140,
  cex_points = 0.5,
  cex_names = 0.5,
  ...
)
}
\description{
TODO: write doc
}
