% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnl_model_functions.R
\name{mnl_randomize_design}
\alias{mnl_randomize_design}
\title{Randomize the order of the choice sets and alternatives}
\usage{
mnl_randomize_design(design, seed = NULL)
}
\arguments{
\item{design}{Array of dimensions \code{(q, J, S)}.}

\item{seed}{Integer seed for reproducibility}
}
\value{
Array of dimensions \code{(q, J, S)}.
}
\description{
This function transforms the order of the choice sets and the alternatives in a design array.
}
