% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_model_functions.R
\name{brent_cox_scheffe_gaussian}
\alias{brent_cox_scheffe_gaussian}
\title{R wrapper for BrentCoxScheffeGaussian in C++.
Minimizes the efficiency function for Scheffé model using Brent's method for local optima.
The function also checks the edge cases (in normal mixtures it's 0 and 1) to see if the minimizers are there.
TODO: write doc}
\usage{
brent_cox_scheffe_gaussian(
  X,
  j,
  i,
  order,
  opt_crit,
  lower = 0,
  upper = 1,
  tol = 1e-04
)
}
\description{
R wrapper for BrentCoxScheffeGaussian in C++.
Minimizes the efficiency function for Scheffé model using Brent's method for local optima.
The function also checks the edge cases (in normal mixtures it's 0 and 1) to see if the minimizers are there.
TODO: write doc
}
