% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_model_functions.R
\name{create_random_initial_design_gaussian}
\alias{create_random_initial_design_gaussian}
\title{TODO: write doc}
\usage{
create_random_initial_design_gaussian(n_runs, q, seed = NULL)
}
\description{
TODO: write doc
}
