% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_model_functions.R
\name{gaussian_create_moment_matrix}
\alias{gaussian_create_moment_matrix}
\title{TODO: write doc}
\usage{
gaussian_create_moment_matrix(q, n_pv = 0, order = 3)
}
\description{
TODO: write doc
}
