% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_model_functions.R
\name{gaussian_create_moment_matrix}
\alias{gaussian_create_moment_matrix}
\title{Computes moment matrix using the linear classical Gaussian model.}
\usage{
gaussian_create_moment_matrix(q, order = 3)
}
\arguments{
\item{q}{Number of mixture ingredients}

\item{order}{integer corresponding to a Scheffé model order (1, 2, or 3).}
}
\description{
Computes moment matrix using the linear classical Gaussian model.
}
