% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_model_functions.R
\name{gaussian_create_moment_matrix}
\alias{gaussian_create_moment_matrix}
\title{Computes moment matrix using the linear classical Gaussian model.}
\usage{
gaussian_create_moment_matrix(q, n_pv = 0, order = 3, pv_bounds = NULL)
}
\arguments{
\item{q}{Number of mixture ingredients}

\item{n_pv}{Number of process variables}

\item{order}{integer corresponding to a Scheffé model order (1, 2, 3 if no process variables are used; 4 if process variables are used).}

\item{pv_bounds}{Vector or matrix that specifies the bounds of the process variables. If NULL, it defaults to the \code{[-1, 1]} interval.}
}
\description{
Computes moment matrix using the linear classical Gaussian model.
}
\details{
\code{pv_bounds} must be either: \itemize{
\item NULL (for default values), or
\item a vector with 2 scalars (the second element must be greater than the first), or
\item a matrix with n_pv rows and 2 columns, i.e., each row represents a bound for each process variable.
}
}
