% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_model_functions.R
\name{gaussian_create_random_initial_design}
\alias{gaussian_create_random_initial_design}
\title{Creation of a random initial design using the linear classical Gaussian model.}
\usage{
gaussian_create_random_initial_design(n_runs, q, seed = NULL)
}
\arguments{
\item{n_runs}{Number of runs in the design.}

\item{q}{Number of mixture ingredients.}

\item{seed}{integer used for reproducibility}
}
\value{
Matrix of size \code{(n_runs, q)}.
}
\description{
Creation of a random initial design using the linear classical Gaussian model.
}
\examples{
gaussian_create_random_initial_design(10, 3, 2, seed = 3)


}
