% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_model_functions.R
\name{gaussian_create_random_initial_design}
\alias{gaussian_create_random_initial_design}
\title{TODO: write doc}
\usage{
gaussian_create_random_initial_design(
  n_runs,
  q,
  n_pv = 0,
  pv_bounds = NULL,
  seed = NULL
)
}
\description{
TODO: write doc
}
