% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_model_functions.R
\name{gaussian_get_opt_crit_value}
\alias{gaussian_get_opt_crit_value}
\title{TODO: write doc}
\usage{
gaussian_get_opt_crit_value(X, order = 1, opt_crit = 0, n_pv = 0)
}
\description{
TODO: write doc
}
