% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{get_halton_draws}
\alias{get_halton_draws}
\title{TODO: write doc
Returns a matrix of dimension m x ndraws where m is the length of the beta vector.
Example on how to use:
beta_means = mnl_create_random_beta(q)
beta_prior_draws = get_halton_draws(beta_means$beta)}
\usage{
get_halton_draws(beta, sd = 1, ndraws = 120)
}
\description{
TODO: write doc
Returns a matrix of dimension m x ndraws where m is the length of the beta vector.
Example on how to use:
beta_means = mnl_create_random_beta(q)
beta_prior_draws = get_halton_draws(beta_means$beta)
}
