% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{mixture_coord_ex}
\alias{mixture_coord_ex}
\title{TODO: write doc}
\usage{
mixture_coord_ex(
  X,
  order = NULL,
  beta = NULL,
  model = "Gaussian",
  n_cox_points = 100,
  max_it = 50,
  plot_designs = F,
  verbose = 1
)
}
\description{
TODO: write doc
}
