% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_model_functions.R
\name{mixture_coord_ex_gaussian}
\alias{mixture_coord_ex_gaussian}
\title{TODO: write doc}
\usage{
mixture_coord_ex_gaussian(
  X,
  order = 1,
  n_cox_points = 100,
  max_it = 50,
  plot_designs = F,
  verbose = 1
)
}
\description{
TODO: write doc
}
