% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnl_model_functions.R
\name{mnl_create_moment_matrix}
\alias{mnl_create_moment_matrix}
\title{Computes moment matrix using the MNL model.}
\usage{
mnl_create_moment_matrix(q, order = 3)
}
\arguments{
\item{q}{Number of mixture ingredients.}

\item{order}{integer corresponding to a Scheffé model order (1, 2, 3).}
}
\value{
Matrix of size \code{(m, m)}.
}
\description{
Computes moment matrix using the MNL model.
}
