% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnl_model_functions.R
\name{mnl_create_random_initial_design}
\alias{mnl_create_random_initial_design}
\title{Creation of a random initial design using the MNL model}
\usage{
mnl_create_random_initial_design(q, J, S, seed = NULL)
}
\arguments{
\item{q}{integer specifying the number of mixture ingredients}

\item{J}{integer specifying the number of alternatives within each choice set}

\item{S}{integer specifying the number of choice sets}

\item{seed}{integer used for reproducibility}
}
\value{
3-dimensional array of size \code{(q, J, S)}.
}
\description{
Creates a random initial design using the MNL model of the specified dimensions.
}
\examples{
mnl_create_random_initial_design(q = 3, J = 2, S = 8, seed = 2020)


}
