% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnl_model_functions.R
\name{mnl_design_array_to_dataframe}
\alias{mnl_design_array_to_dataframe}
\title{TODO: write doc
Converts a design in an array to a more readable or plotable data frame.
des_array must be a 3-dimensional design array of dimension (k, J, S) where k is the number of variables, J the number of alternatives in each choice set and S is the number of choice sets.
Returns a dataframe  with (k+1) columns where the first q columns are the variables and the (k+1)-th column is the choice set.
If names is null then the names of the columns are c(paste0("c", 1:k), "choice_set").
Example:
des_array = mnl_create_random_initial_design(3, 2, 4)
design_array_to_dataframe(des_array, names = c("v1", "v2", "v3", "choice_set"))}
\usage{
mnl_design_array_to_dataframe(des_array, names = NULL)
}
\description{
TODO: write doc
Converts a design in an array to a more readable or plotable data frame.
des_array must be a 3-dimensional design array of dimension (k, J, S) where k is the number of variables, J the number of alternatives in each choice set and S is the number of choice sets.
Returns a dataframe  with (k+1) columns where the first q columns are the variables and the (k+1)-th column is the choice set.
If names is null then the names of the columns are c(paste0("c", 1:k), "choice_set").
Example:
des_array = mnl_create_random_initial_design(3, 2, 4)
design_array_to_dataframe(des_array, names = c("v1", "v2", "v3", "choice_set"))
}
