% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnl_model_functions.R
\name{mnl_design_dataframe_to_array}
\alias{mnl_design_dataframe_to_array}
\title{TODO: write doc
Returns a 3 dimensional array to be used in the functions in the package
des_df must be a dataframe  with (k+1) columns where the first k columns are the variables and the (k+1)-th column is the choice set.}
\usage{
mnl_design_dataframe_to_array(des_df)
}
\description{
TODO: write doc
Returns a 3 dimensional array to be used in the functions in the package
des_df must be a dataframe  with (k+1) columns where the first k columns are the variables and the (k+1)-th column is the choice set.
}
