% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnl_model_functions.R
\name{mnl_get_fds_simulations}
\alias{mnl_get_fds_simulations}
\title{TODO: write doc
Returns a dataframe with Monte Carlo simulations for Fraction of Design Space plots
Output dataframe is of size J*n_points_per_alternative where J is the number of alternatives per choice set, i.e., the second element in dim(design_array).}
\usage{
mnl_get_fds_simulations(
  design_array,
  beta,
  order,
  n_points_per_alternative = 500,
  transform_beta = F
)
}
\description{
TODO: write doc
Returns a dataframe with Monte Carlo simulations for Fraction of Design Space plots
Output dataframe is of size J*n_points_per_alternative where J is the number of alternatives per choice set, i.e., the second element in dim(design_array).
}
