% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnl_model_functions.R
\name{mnl_get_fds_simulations}
\alias{mnl_get_fds_simulations}
\title{Perform Monte Carlo simulations for FDS plots.}
\usage{
mnl_get_fds_simulations(
  design_array,
  beta,
  order,
  n_points_per_alternative = 500,
  transform_beta = F,
  verbose = 0
)
}
\arguments{
\item{design_array}{3-dimensional design array.}

\item{beta}{Matrix with Halton draws of the prior distribution}

\item{order}{integer corresponding to Scheffé model of order 1, 2, or 3.}

\item{n_points_per_alternative}{Number of points to approximate the prediction variance.}

\item{transform_beta}{boolean parameter. Should the beta vector/matrix be transformed by subtracting the \code{q}-th element?}

\item{verbose}{Level of verbosity}
}
\value{
Tibble of size \code{J*n_points_per_alternative} where J is the number of alternatives per choice set, i.e., the second element in \code{dim(design_array)}.
}
\description{
This function creates a dataframe/tibble with Monte Carlo simulations for Fraction of Design Space plots.
}
