% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnl_model_functions.R
\name{mnl_get_information_matrix}
\alias{mnl_get_information_matrix}
\title{TODO: write doc
Gets the information matrix of the MNL model.
Parameters:
X: 3-dimensional array
beta: vector
order
transform_beta}
\usage{
mnl_get_information_matrix(X, beta, order, transform_beta)
}
\description{
TODO: write doc
Gets the information matrix of the MNL model.
Parameters:
X: 3-dimensional array
beta: vector
order
transform_beta
}
