% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnl_model_functions.R
\name{mnl_get_opt_crit_value}
\alias{mnl_get_opt_crit_value}
\title{Function that computes optimality criterion value for MNL model}
\usage{
mnl_get_opt_crit_value(
  X,
  beta,
  order,
  opt_crit = "D",
  transform_beta = T,
  n_pv = 0
)
}
\arguments{
\item{X}{3 dimensional array with dimensions (q, J, S) where:
q is the number of ingredient proportions,
J is the number of alternatives within a choice set,
S is the number of choice sets.}

\item{beta}{numeric vector containing the parameters or numeric matrix containing draws of the prior distribution of the parameters.}

\item{opt_crit}{optimality criterion: 0 or "D" is D-optimality and 1 or "I" is I-optimality}
}
\value{
Returns the value of the optimality criterion for this particular design and this beta vector
}
\description{
\code{mnl_get_opt_crit_value} computes optimality criterion value for MNL model
}
