% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnl_model_functions.R
\name{mnl_get_Ps}
\alias{mnl_get_Ps}
\title{Compute choice probabilities of a specific choice set using the MNL model.}
\usage{
mnl_get_Ps(X, beta, s, order = 3, transform_beta = T)
}
\arguments{
\item{X}{3-dimensional design cube/array of size \code{(q, J, S)}.}

\item{beta}{Parameter vector}

\item{s}{integer corresponding to a choice set in 1 to S.}

\item{order}{integer corresponding to a Scheffé model order (1, 2, 3).}

\item{transform_beta}{boolean parameter. Should the beta vector/matrix be transformed by subtracting the \code{q}-th element?}
}
\value{
J-dimensional numerical vector.
}
\description{
Compute choice probabilities of a specific choice set using the MNL model.
}
